Fixed decryption and rewrote the face mesh
This code is a facetracker that uses AES CTR mode to encrypt and decrypt using landmarks and key features of faces by people. Then uses these attributes to encrypt a message in AES CTR mode. I took Kash's code and rewrote it so that it could work on my computer the face mesh function is different as it uses media pipe and not dlib. Also, the decryption is fixed allowing a user to encrypt and decrypt a message properly.